+++ Changes.md
@@ -22,6 +22,7 @@ API
   - When setting the value, an OS-native or generic string can be supplied. It will automatically be converted to Gaffer's internal file path representation.
   - Values returned by `getValue()` are in OS-native format.
   - `FilePathPlug` is compatible with `StringPlug` - `FilePathPlug` can be used as input for `StringPlug` and `StringPlug` can be used as input for `FilePathPlug`.
+  - Converted existing `StringPlug` used for file paths to `FilePathPlug`.
 
 Breaking Changes
 ----------------